Move main test annotation processor to test package
